+++ test/unit/translations/maps/linkedPhysicalIndexToValueMap.spec.js
@@ -59,6 +59,25 @@ it('should work with get, and set functions properly', () => {
   expect(indexToValueMap.getLength()).toBe(3);
 });
 
+it('should work with set function properly (working with the "position" argument)', () => {
+  const indexToValueMap = new IndexToValueMap();
+
+  indexToValueMap.init(5);
+  indexToValueMap.setValueAtIndex(0, 2);
+  indexToValueMap.setValueAtIndex(2, 0);
+  indexToValueMap.setValueAtIndex(1, 1);
+
+  expect(indexToValueMap.orderOfIndexes).toEqual([0, 2, 1]);
+
+  indexToValueMap.setValueAtIndex(4, 3, 0);
+
+  expect(indexToValueMap.orderOfIndexes).toEqual([4, 0, 2, 1]);
+
+  indexToValueMap.setValueAtIndex(3, 4, 2);
+
+  expect(indexToValueMap.orderOfIndexes).toEqual([4, 0, 3, 2, 1]);
+});
+
 it('should init map properly when passing function as initialization property', () => {
   const indexToValueMap = new IndexToValueMap(index => ({ key: index }));
 
@@ -91,6 +110,28 @@ it('should clear values properly', () => {
   expect(indexToValueMap.getValues()).toEqual([]);
   expect(indexToValueMap.getLength()).toBe(0);
   expect(indexToValueMap.orderOfIndexes).toEqual([]);
+});
+
+it('should clear single value properly', () => {
+  const indexToValueMap = new IndexToValueMap();
+
+  indexToValueMap.init(3);
+  indexToValueMap.setValues([{ key: 1 }, { key: 2 }, { key: 0 }]);
+  expect(indexToValueMap.orderOfIndexes).toEqual([0, 1, 2]);
+
+  indexToValueMap.clearValue(1);
+
+  expect(indexToValueMap.indexedValues).toEqual([{ key: 1 }, null, { key: 0 }]);
+  expect(indexToValueMap.getValues()).toEqual([{ key: 1 }, { key: 0 }]);
+  expect(indexToValueMap.getLength()).toBe(2);
+  expect(indexToValueMap.orderOfIndexes).toEqual([0, 2]);
+
+  indexToValueMap.clearValue(0);
+
+  expect(indexToValueMap.indexedValues).toEqual([null, null, { key: 0 }]);
+  expect(indexToValueMap.getValues()).toEqual([{ key: 0 }]);
+  expect(indexToValueMap.getLength()).toBe(1);
+  expect(indexToValueMap.orderOfIndexes).toEqual([2]);
 });
 
 it('should handle `insert` method properly', () => {